Make too many changes to summarize
It's been far too long since I made a commit to this repo.  There are
two main overarching changes along with a myriad of smaller ones: the
libraries have been broken out into seperate packages and the entire
repository has been rearranged.

All libraries are now stored in lib/ and are installed in venv's
site-packages directory via symlinks.  I know this is a super hacky way
to do it and a better solution will come in the future when the
libraries are more mature and possibly published to PyPI.  For now, a
convenience script is provided which will be updated as needed
(create_library_links.sh).

One of the "smaller" changes is the introduction of cache_util/.  This
will contain any scripts that are there to manage a local xkcd comic
and metadata cache.  Utilities will be added to both manage the cache
better and test features in the xkcd library being developed alongside
the discord bot.

Another small change that is actually relevant to the discord bot is a
bug fix where the gateway would close the socket with an "unable to
parse packet" error.  This was being caused by the heartbeat packets
sending "d: null" if a packet sequence hadn't been recieved yet.  The
fix for this was to change GatewayPayload.to_json to only add members
that are not None to the resulting json string.

The next step for this project should be to move all json-repersenting
classes to the new json_obj decorator and flushing out the xkcd
library.